Lightweight STM32 serial terminal in C
(This is not a solution to Issue #4, which will provide a GUI interface for testing, etc.)

For those interested in a snappy terminal-based serial interface to the STM32 development board. Source code shows some low-level serial port IO and multiple terminal windows in ncurses.

Works on macos, probably works out of the box in linux, not sure about windows.

CTS1+ is prefixed to telecommands, and the serial protocol is automatically set.