In-memory LOG storage for most recent messages
Logs are stored in a circular buffer.
Allows limited recursion of LOG_message().
Telecommands for reporting the latest and the N-latest log
messages from in-memory log table.

The in-memory log table serves as statically-allocated log message storage that
allows concurrent and nested calls to LOG_message() and serves as a limited
backup for the filesystem logs. This is not considered a traditional log sink.