Added python script to process and read camera files
script takes text as its outputted on the ground station UI, removes log messages, and turns image data into a jpg image.
NOTE in its current state the script removes any empty line, and any line that has a colon `:` in it, and removes `TODO` as well as a set number of characters after it. The script then merges separated data into the previous line to ensure all lines start with `@` and are a fixed length before creating the image